requirements.txt
xgboost==1.6.2
numpy==1.23.3
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
matplotlib==3.6.0
jupyterlab==3.4.7
Pillow==9.2.0
pandas==1.5.0
opencv-python==4.6.0.66
scipy==1.9.1
tqdm==4.64.1
scikit-learn==1.1.2
